fix: Low-Power Assumptions
This test assumed the pin dedicated to USER_BTN was pulled high
by the hardware. Now an internal pull-up is enabled to ensure
this behavior. The custom ST input mode of `INPUT_ANALOG` is now
used on the GPIO to ensure the input pins are floating and not
sinking any exteraneous current.